moved to py12 | started refactor | broke everything into multiple files | updated person admin, custom forms
