+++ .gitignore
@@ -5,3 +5,4 @@ org_data/*
 static/
 data/
 fly-pg-dump-to-s3/
+node_modules/
