+++ app/admin/person.py
@@ -1,0 +1,273 @@
+from django.contrib import admin
+from django.forms import ModelForm, Textarea, ValidationError, fields
+from django.urls import reverse
+from django.utils.html import format_html
+from import_export import resources, widgets
+from import_export.admin import ImportExportModelAdmin
+from import_export.fields import Field
+
+from app.models.person import Person
+from app.models.prison import PersonPrison, Prison
+from app.utils import WorkflowStage
+
+
+class PersonResource(resources.ModelResource):
+    current_prison = Field(attribute="current_prison")
+    last_served = Field(attribute="last_served")
+    eligibility = Field(attribute="eligibility")
+    package_count = Field(attribute="package_count")
+    letter_count = Field(attribute="letter_count")
+    # this causes legacy_last_served_date to be part of the export
+    # even though it is probably confusing
+    legacy_last_served_date = Field(
+        attribute="legacy_last_served_date",
+        column_name="legacy_last_served_date",
+        widget=widgets.DateTimeWidget(format="%Y-%m-%d %H:%M:%S"),
+    )
+
+    class Meta:
+        model = Person
+        skip_unchanged = True
+        report_skipped = True
+        fields = (
+            "id",
+            "inmate_number",
+            "last_name",
+            "middle_name",
+            "first_name",
+            "name_suffix",
+            "status",
+            "legacy_last_served_date",
+            "legacy_prison_id",
+        )
+        readonly_fields = (
+            "current_prison",
+            "last_served",
+            "eligibility",
+            "letter_count",
+            "package_count",
+        )
+        export_order = (
+            "id",
+            "last_name",
+            "middle_name",
+            "first_name",
+            "name_suffix",
+            "status",
+            "current_prison",
+            "last_served",
+            "eligibility",
+            "letter_count",
+            "package_count",
+        )
+
+    def after_save_instance(self, instance, using_transactions, dry_run):
+        if instance.legacy_prison_id is None:
+            return
+        legacy_prison_id = instance.legacy_prison_id
+        prison = Prison.objects.filter(legacy_id=legacy_prison_id).first()
+        if not prison:
+            raise Exception(f"Prison with legacy id {legacy_prison_id} not found!")
+        existing_records = PersonPrison.objects.filter(person_id=instance.id).count()
+        if existing_records > 0:
+            return
+        PersonPrison.objects.create(person_id=instance.id, prison_id=prison.id)
+
+
+class PersonAdminForm(ModelForm):
+    allow_empty_inmate = False
+
+    class Meta:
+        model = Person
+        fields = [
+            "inmate_number",
+            "last_name",
+            "middle_name",
+            "first_name",
+            "name_suffix",
+            "status",
+            "notes",
+        ]
+
+    def clean_inmate_number(self):
+        inmate_number = self.cleaned_data.get("inmate_number", "")
+        if not self.allow_empty_inmate and not inmate_number:
+            raise ValidationError("Inmate ID is required")
+        inmate_number = "".join(list(filter(str.isalnum, inmate_number)))
+        return inmate_number.upper()
+
+    def clean_first_name(self):
+        return self.cleaned_data["first_name"].upper()
+
+    def clean_middle_name(self):
+        return self.cleaned_data["middle_name"].upper()
+
+    def clean_last_name(self):
+        return self.cleaned_data["last_name"].upper()
+
+    def clean_name_suffix(self):
+        return self.cleaned_data["name_suffix"].upper()
+
+
+class PersonCreateForm(PersonAdminForm):
+    inmate_number = fields.CharField(required=False)
+    did_not_include_inmate_number = fields.BooleanField(initial=False, required=False)
+
+    allow_empty_inmate = True
+
+    class Meta(PersonAdminForm.Meta):
+        fields = [
+            "inmate_number",
+            "did_not_include_inmate_number",
+            "last_name",
+            "middle_name",
+            "first_name",
+            "name_suffix",
+            "status",
+            "notes",
+        ]
+
+    def clean(self):
+        cleaned_data = super().clean()
+        inmate_number = cleaned_data.get("inmate_number", "")
+        did_not = cleaned_data.get("did_not_include_inmate_number", False)
+        if not inmate_number and not did_not:
+            raise ValidationError(
+                "No inmate number included. If one was not provided, leave blank but also check 'Did Not Include Inmate Number'"
+            )
+        elif inmate_number and did_not:
+            raise ValidationError(
+                "Inmate ID must be blank if 'Did Not Include Inmate Number' is checked"
+            )
+        elif did_not:
+            if not (last_person := Person.objects.all().order_by("pk").last()):
+                last_pk = 0
+            else:
+                last_pk = last_person.pk
+            cleaned_data["inmate_number"] = f"UNKNOWNID{last_pk}"
+        return cleaned_data
+
+
+class PersonPrisonInline(admin.TabularInline):
+    model = PersonPrison
+    max_num = 1
+    can_delete = False
+    verbose_name = "Prison"
+    verbose_name_plural = "Prisons"
+    fields = ("prison",)
+
+    def formfield_for_foreignkey(self, db_field, request, **kwargs):
+        if db_field.name == "prison":
+            kwargs["queryset"] = Prison.objects.all().order_by("name")
+        return super().formfield_for_foreignkey(db_field, request, **kwargs)
+
+    def get_formset(self, request, obj=None, **kwargs):
+        """
+        Override the formset function in order to remove the add and
+        change buttons beside the foreign key pull-down menus in the inline.
+        From: https://stackoverflow.com/a/37558444
+        """
+        formset = super().get_formset(request, obj, **kwargs)
+        form = formset.form
+        widget = form.base_fields["prison"].widget
+        widget.can_add_related = False
+        widget.can_change_related = False
+        return formset
+
+
+# TODO REFACTOR: create form needs "no id sent" checkbox which allows null person_id field
+# and auto generates one for them
+class PersonAdmin(ImportExportModelAdmin):
+    resource_class = PersonResource
+
+    def last_served_date(self, obj) -> str | None:
+        if obj.last_served:
+            return obj.last_served.strftime("%Y-%m-%d")
+
+    list_display = (
+        "inmate_number",
+        "last_name",
+        "first_name",
+        "eligibility",
+        "status",
+        "last_served_date",
+        "current_prison",
+        "package_count",
+        "pending_letter_count",
+        "letter_count",
+        "created_by",
+        "created_date",
+        "modified_date",
+    )
+    readonly_fields = (
+        "current_prison",
+        "created_by",
+        "created_date",
+        "modified_date",
+        "eligibility",
+        "package_count",
+        "pending_letter_count",
+        "letter_count",
+    )
+    list_filter = (
+        # "prisons__prison",
+        # EligibilityListFilter,
+    )
+    search_fields = (
+        "inmate_number",
+        "last_name",
+        "first_name",
+    )
+    list_display_links = (
+        "first_name",
+        "last_name",
+    )
+    list_per_page = 25
+    inlines = [PersonPrisonInline]
+
+    def get_form(self, request, obj=None, **kwargs):
+        if not obj:
+            return PersonCreateForm
+        return PersonAdminForm
+
+    def get_fields(self, request, obj=None):
+        if obj:
+            return PersonAdminForm.Meta.fields
+        else:
+            return PersonCreateForm.Meta.fields
+
+    def current_prison(self, person):
+        if not person.current_prison:
+            return
+        link = reverse(
+            "admin:app_prison_change", kwargs={"object_id": person.current_prison.id}
+        )
+        return format_html(f"<a href={link}>{person.current_prison.name}</a>")
+
+    current_prison.allow_tags = True
+
+    def pending_letter_count(self, person):
+        if not person.pending_letter_count:
+            return
+        return format_html(
+            f"<a href={reverse('admin:app_letter_changelist')}?person={person.id}&workflow_stage__in={WorkflowStage.STAGE1_COMPLETE} target='_blank'>{person.pending_letter_count}</a>"
+        )
+
+    def letter_count(self, person):
+        if not person.letter_count:
+            return
+        return format_html(
+            f"<a href={reverse('admin:app_letter_changelist')}?person={person.id}>{person.letter_count}</a>"
+        )
+
+    def package_count(self, person):
+        if not person.package_count:
+            return
+        return format_html(
+            f"<a href={reverse('admin:app_letter_changelist')}?person={person.id}&workflow_stage__in={WorkflowStage.FULFILLED}>{person.package_count}</a>"
+        )
+
+    def save_model(self, request, obj, form, change):
+        if not obj.pk:
+            obj.created_by = request.user
+        super().save_model(request, obj, form, change)
